12/04/2023 Codes
2.	Re-wite the above program to find a pattern of letters in the set lines read from 
a.	User input (maximum 10 lines to be read)
b.	Files
Use user-defined functions and display the lines having the matched pattern of letters
Do not use any inbuilt function. 

Ah, Dear! could you and I with Fate conspire
	To grasp this sorry Scheme of Things entire,
	Would not we shatter it to bits -- and then 
	Re-mould it nearer to the Heart’s Desire

Pattern: “ould”
	
Ah, Dear! could you and I with Fate conspire
Would not we shatter it to bits -- and then 
Re-mould it nearer to the Heart’s Desire