Dijkstras Algorithm(Graphs)
 Dijkstra's algorithm:

Initialize the algorithm by selecting a start node and assigning it a distance of 0. Set the distance of all other nodes to infinity.

Mark all nodes as unvisited. Create a priority queue or a min-heap to store the nodes and their corresponding distances.

While there are unvisited nodes, repeat the following steps:

Select the node with the smallest distance from the priority queue.
Mark the selected node as visited.
For each neighbor of the selected node that is still unvisited, calculate the distance from the start node through the selected node. If this distance is smaller than the neighbor's current distance, update the neighbor's distance.

After updating the distances of all neighbors, enqueue them into the priority queue.

Repeat steps 3-5 until all nodes have been visited or until the destination node is reached.

Once the destination node is reached or all nodes have been visited, the algorithm terminates. The shortest path from the start node to the destination node can be obtained by backtracking from the destination node, following the nodes with the lowest distances.